change return type of getOsTid() const; from long to IntPtr Signed-off-by: FrancisANDRE <zosrothko@orange.fr>

+++ Foundation/include/Poco/Message.h
@@ -128,7 +128,7 @@ public:
 	long getTid() const;
 		/// Returns the numeric thread identifier for the message.
 	
-	long getOsTid() const;
+	IntPtr getOsTid() const;
 		/// Returns the numeric OS thread identifier for the message.
 	
 	void setPid(long pid);
